requirements.txt
firebase_admin
yt_dlp
python-dotenv
flask
pygame
pyserial
scipy
numpy
requests